serial: 8250: Add cpufreq support
On DA850/OMAP-L138 SoC, the PLL which supplies the clock to CPU also
feeds the
UART and the UART input frequency can change when the CPU frequency is
scaled.

This patch adds cpufreq support for 8250 serial driver. A clk structure
member
has been added to the platform and port data structures. This member is
used by
the cpufreq notifier callback to get the updated clock rate. The
implementation
is based on the cpufreq implementation for Samsung serial driver.

Tested on TI DA850/OMAP-L138 EVM.

Signed-off-by: Chaithrika U S <chaithrika@ti.com>