da850/omap-l138: Configure GP0[11] only when UI card is detected
On Logic PD Rev.3 EVMs GP0[11] pin needs to be configured
for MMC and NOR to work. When GP0[11] is low, the SD0
interface will not work, but NOR flash will. When GP0[11]
is high, SD0 will work but NOR flash will not. By default
we are assuming that GP0[11] pin is driven high, when UI
card is not connected. Hence we are not configuring the
GP0[11] pin when MMC/SD is enabled and UI card is not
connected. Not configuring the GPIO pin will enable the
bluetooth to work on AM18x as it requires the GP0[11]
pin for UART flow control.

Signed-off-by: Rajashekhara, Sudhakar <sudhakar.raj@ti.com>