Revert "ASoC: Configure symmetric rates for tlv320aic3x"
This reverts commit 140176159597ea1f23dcccb47b5c38fdf7c7faa8.
This commit configures the symmetric rates for tlc320aic3x codec. But
during debugging, it was found that whether "symmetric_rates" is set or
not, the sampling rate was getting modified.

Additionally, with this commit, "arecord | aplay" was fails when
executed just after the board boots. Till a better fix is found, revert
this commit.