davinci: vpfe: previewer driver based on v4l2 media controller framework
Add the video previewer driver with the v4l2 media controller framework
which takes care of converting the video frames from bayer to YUV
or RGB. the driver supports both continuous mode where it works in
tandem with the CCDC and single shot mode where it can be used in
isolation. The driver underneath uses the dm365 IPIPE module for
programming the hardware.
The driver supports previewer as a subdevice and a media entity,
and the enumerable pads are 2(1 input and 1 output). The driver
supports streaing and all the pad and link related operations.
Specific functions like defect pixel correction, LUT are supported
through private ioctls.

Signed-off-by: Nagabhushana Netagunte <nagabhushana.netagunte@ti.com>