media: add support for decoder subdevs along with sensor and others
A lot of SOCs including Texas Instruments Davinci family mainly use
video decoders as input devices. Here the initial subdevice node
from where the input really comes is this decoder, for which support
is needed as part of the Media Controller infrastructure. Ths patch
adds an additional flag to include the decoders along with others,
such as the sensor and lens.

Signed-off-by: Nagabhushana Netagunte <nagabhushana.netagunte@ti.com>